print/foomatic4-db-engine: Fix man page generation on DragonFly(+)
The perl man page generation script has an unnecessary file type check
that causes no man pages to be generated on DragonFly and possibly other
platforms.

After properly identifying the template file, it checks to see if it's type
nroff, troff, or groff.  On DragonFly it is type "ASCII".  Just skip the
check completely, the templates were extracted from the verified tarball.