Initial import of clusterssh, version 4.01.01, into the NetBSD Packages Collection.
The is the Perl application bundle for ClusterSSH (a.k.a cssh), formally
a GNU tools based project.

ClusterSSH is a tool for making the same change on multiple servers at
the same time.  The 'cssh' command opens an administration console and
an xterm to all specified hosts.  Any text typed into the administration
console is replicated to all windows.  All windows may also be typed into
directly.

This tool is intended for (but not limited to) cluster administration
where the same configuration or commands must be run on each node
within the cluster.  Performing these commands all at once via this
tool ensures all nodes are kept in sync.